requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
httpx
pydantic-settings
requests
beautifulsoup4